Hooking the Network manager up to the ProfilerNotifier
