+++ com.unity.multiplayer.mlapi/Runtime/Profiling/IHasProfilableTransport.cs
@@ -1,0 +1,7 @@
+namespace MLAPI.Profiling
+{
+    public interface IHasProfilableTransport
+    {
+        ITransportProfilerData GetTransport();
+    }
+}
